% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singer.R-package.R
\docType{package}
\name{singer.R-package}
\alias{singer.R}
\alias{singer.R-package}
\title{singer.R: Singer for R}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Singer is an open-source standard for writing scripts that move data. This package allows you to orchestrate Singer taps and targets from R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tjpalanca/singer.R}
  \item Report bugs at \url{https://github.com/tjpalanca/singer.R/issues}
}

}
\author{
\strong{Maintainer}: Troy James Palanca \email{mail@tjpalanca.com} (\href{https://orcid.org/0000-0002-0610-9213}{ORCID})

}
\keyword{internal}
